Update SQF/dayz_code/actions/refuel.sqf
This appears to be a bug. I found [this file](https://github.com/R4Z0R49/DayZMod/blob/1.7.5/SQF/dayz_code/compile/local_setFuel.sqf) and I'm assuming the names need to match in order to actually set the vehicle Fuel level.

I am trying to familiarize myself with the SQF syntax and in trying to do so I came across this. I would appreciate if R4Z0R49 or UnclearWall could maybe give me a quick primer on the gist of what goes on in a script such as this. 

Also: Can you explain how to test a fix like this locally? As in how would I check if this breaks something? 